requirements.txt
locust==2.20.2
python-telegram-bot==20.3
Pillow==10.1.0
requests==2.31.0
numpy==1.26.0
torch==2.1.0 --index-url https://download.pytorch.org/whl/cu118
setuptools==68.2.2